+++ programs/midas-vault/src/state/vault_common_state.rs
@@ -17,10 +17,10 @@ pub struct VaultCommonState {
     pub tokens_receiver: Pubkey,
     pub fee_receiver: Pubkey,
     pub instant_fee: u64,
-    pub instant_daily_limit: u64,
+    pub instant_daily_limit: u128,
     pub variation_tolerance: u64,
     pub min_amount: u64,
 
     pub instant_last_day: u32,
-    pub instant_daily_limit_used: u64,
+    pub instant_daily_limit_used: u128,
 }
